requirements.txt
requests
supabase
python-dotenv
openai
